go: update to 1.17.7
go1.17.7 (released 2022-02-10) includes security fixes to the
crypto/elliptic, math/big packages and to the go command, as well as bug
fixes to the compiler, linker, runtime, the go command, and the
debug/macho, debug/pe, and net/http/httptest packages. See the Go 1.17.7
milestone on our issue tracker for details.

https://github.com/golang/go/issues?q=milestone%3AGo1.17.7+label%3ACherryPickApproved