Cleanup decoder/encoder stuff for L4T Builds, update some things, fix GLES building, but keep GL build. Other Minor cleanups. (#1623)
* Update L4T encoder/decoder to latest version which is 70% faster at 1080p

* Add vulkan support to ffmpeg

* Add vitaquake3 and boom3 cores for switch builds

These cores are considered broken, but starting, so maybe in the future they will be more reliable.

* Run dock-hotplug before starting retroarch.

This Will properly force DPI if booted on dock, and audio switching. It also has the added bonus of force fixing audio issues where speakers randomly dont work on boot."

* Update Kernel Repos

* update libglvnd

* update xf86-video-nouveau

* Build mesa properly in L4T LibreELEC builds.

* Update L4T build config for retroarch to fix building with support for EGL/openGLES.
While L4T supports EGL/OpenGLES, and the issue building for it is fixed,
we arent switching now though because moonlight core doesnt support it.

* Exclude moonlight core in Lakka builds that use openGLES over openGL, as openGLES is not supported

* Use openGLES in LibreELEC builds, as that is what upstream is targeting