bluez: update to 5.64
Release of BlueZ 5.64
- 19th March 2022, 06:18 am by Tedd Ho-Jeong An

This is another release mostly with the bug fixes on HOG, GATT, A2DP,
Media, AVDTP, AVRCP, and scanning failure. Also, this release includes a
fix for building with old glibc (< 2.25), and other minor issues found
with the static code analyzing tool. ISO packet support is added to the
emulator as a part of LE Audio development.