linux: fix oldconfig handling
make oldconfig is very noisy when config options are missing and
produces a lot of errors as it's trying to read from stdin which
is not available during build.

Fix this by piping yes "" to it's stdin to select default kernel
options.

Also run "make listnewconfig" before that to get a summary of the
additional options set like it's done eg in Debian.

Signed-off-by: Matthias Reichl <hias@horus.com>