gcc: update to 12.2.0
This is the list of problem reports (PRs) from GCC's bug tracking system
that are known to be fixed in the 12.2 release. This list might not be
complete (that is, it is possible that some PRs that have been fixed are
not listed here).

https://gcc.gnu.org/bugzilla/buglist.cgi?bug_status=RESOLVED&resolution=FIXED&target_milestone=12.2

Target Specific Changes

LoongArch

The default setting of -m[check|no-check]-zero-division is changed for
optimized code. Now -mno-check-zero-division is the default for all
optimization levels but -O0 and -Og. The old behavior can be obtained by
explicitly passing -mcheck-zero-division to GCC.