ffmpeg: update patch handling
Use upstream ffmpeg as a base version and apply kodi patches via a
patchset.

This doesn't change the ffmpeg source code we build from ATM but
allows us to easily drop kodi patches if they cause issues, update
to a newer point release or test with different ffmpeg versions
in the future.

Signed-off-by: Matthias Reichl <hias@horus.com>