update_retroplayer-addons: temporarily exclude game.libretro
The Matrix branch of kodi game repo currently switches game.libretro
from Matrix to retroplayer branch which results in flipping back
game.libretro from 19.0.3 (the latest version in the Matrix branch)
to 19.0.2 as that is the latest tag in that branch.

Exclude it from automatic updates and let's update it manually for now
to avoid those flip-backs.

Signed-off-by: Matthias Reichl <hias@horus.com>