busybox: fs-resize: create exFAT partition
many users request a partition, that is also accessible on operating
system, that does not natively support ext4 system, to be able transfer
their game files to the storage device (see #646).

by reducing the size of the ext4 partition (which mostly stores
configuration and setting files) to roughly 2GiB, the remaining space of
the storage device can be used as exFAT partition, which is accessible
on most (if not all) operating systems.

the partition is mounted by udevil to /storage/roms/EXFAT.

if for some reason the size of the first (FAT32) partition will be
increased in the future, then also the value (4GiB) for the end of the
ext4 partition needs to be adjusted.

this does not break existing installations of Lakka, but it also does
not add the partition to existing installation of Lakka. the partition
is created only on the first boot of the freshly flashed image.