scripts/build: properly honor strip build flag
Package building honor's the strip flag so we can prevent executables
from being stripped.

But copying files to INSTALL ignored that flag which meant the flag
had practically no effect.

Fix this by using the same logic as we do in the meson strip handling in
line 170 and in config/functions.

Also add the missing quotes in scripts/build to align it with config/functions

Signed-off-by: Matthias Reichl <hias@horus.com>