Add upower and libgudev so kodi can get battery stats
Upower was removed from libreELEC in 2013 with this commit: 3a130cd5009c8d3595b8afcde2421324572ce58b
This was a stupid choice as it is the only option in kodi for supplying any battery statitistics. Currently broken.
This patchset: https://github.com/xbmc/xbmc/pull/17327/commits for kodi seemingliy wants to fix the issue(Doesnt, is more of a new way to deal with dbus in general), but will still require that upower be available for battery.